==> scripts/analysis/suppression/4 - Analyze accounts.ipynb <==
# Analyze accounts

In this step we search for insights about the accounts.

import collections
import json
import os

import matplotlib.pyplot as plt
import pandas as pd
from pandas.plotting import register_matplotlib_converters


register_matplotlib_converters()

%matplotlib inline

plt.style.use("seaborn")

## Load the data

data_directory = os.path.join("..", "..", "..", "data")

campaigns_file_path = os.path.join(data_directory, "suppression_campaigns.json")
if not os.path.exists(campaigns_file_path):
    with tarfile.open(os.path.join(data_directory, "suppression_campaigns.tar.xz"), "r", encoding="utf-8") as compressed_file:
        compressed_file.extract("suppression_campaigns.json", data_directory)

with open(campaigns_file_path, "r", encoding="utf-8") as campaigns_file:
    campaigns = []
    line = campaigns_file.readline().strip()
    while line != "":
        campaigns.append(json.loads(line))
        line = campaigns_file.readline().strip()

with open(os.path.join(data_directory, "suppression_bot_hash_to_cluster_id.json"), "r") as json_file:
    bot_hash_to_cluster_id = json.load(json_file)

with open(os.path.join(data_directory, "suppression_attacker_hash_to_cluster_id.json"), "r") as json_file:
    attacker_hash_to_cluster_id = json.load(json_file)

## Collect into data frames

suppressed_contract_names = collections.defaultdict(lambda: set())

# sets
attackers = set()
suppressed_contracts = set()
bots = set()
# campaign counters
attacker_campaign_count = collections.Counter()
suppressed_contract_campaign_count = collections.Counter()
bot_campaign_count = collections.Counter()
# round counters
attacker_round_count = collections.Counter()
suppressed_contract_round_count = collections.Counter()
bot_round_count = collections.Counter()
# transaction counters
attacker_transaction_count = collections.Counter()
suppressed_contract_transaction_count = collections.Counter()
bot_transaction_count = collections.Counter()
# relationships
attacker_bots = collections.defaultdict(lambda: set())
attacker_suppressed_contracts = collections.defaultdict(lambda: set())
bot_attackers = collections.defaultdict(lambda: set())
bot_suppressed_contracts = collections.defaultdict(lambda: set())
suppressed_contract_attackers = collections.defaultdict(lambda: set())
suppressed_contract_bots = collections.defaultdict(lambda: set())
bot_first_blocks = dict()
# profits
attacker_profits = collections.defaultdict(lambda: 0.0)
bot_profits = collections.defaultdict(lambda: 0.0)

for campaign in campaigns:
    # shortcuts
    suppressed_contract = campaign["suppressed_contract_address"]
    bot = campaign["bot_address"]
    rounds = campaign["nr_of_rounds"]
    transactions = campaign["nr_of_transactions"]
    profit = campaign["profit_usd"]
    
    suppressed_contract_names[suppressed_contract] = campaign["suppressed_contract_name"]
    
    # they are never the same entities
    assert suppressed_contract != bot
    
    # update the sets
    suppressed_contracts.add(suppressed_contract)
    bots.add(bot)
    
    # update campaign counters
    suppressed_contract_campaign_count[suppressed_contract] += 1
    bot_campaign_count[bot] += 1
    
    # update round counters
    suppressed_contract_round_count[suppressed_contract] += rounds
    bot_round_count[bot] += rounds
    
    # update transaction counters
    suppressed_contract_transaction_count[suppressed_contract] += transactions
    bot_transaction_count[bot] += transactions
    
    # update the relationships
    suppressed_contract_bots[suppressed_contract].add(bot)
    bot_suppressed_contracts[bot].add(suppressed_contract)
    
    if bot not in bot_first_blocks:
        bot_first_blocks[bot] = campaign["first_block"]
    
    # update the profit
    bot_profits[bot] += profit
    
    # update all for attackers
    manual_transaction_count = 0
    campaign_attackers = set()
    for round_ in campaign["rounds"]:
        round_attackers = set()
        # for transaction attackers
        for transaction in round_["transactions"]:
            assert transaction["to"] == bot
            manual_transaction_count += 1
            
            attacker = transaction["from"]
            attackers.add(attacker)
            campaign_attackers.add(attacker)
            round_attackers.add(attacker)
            
            
            attacker_transaction_count[attacker] += 1
        
        # for round attackers
        for attacker in round_attackers:
            attacker_round_count[attacker] += 1
        
    assert manual_transaction_count == transactions
    
    # for campaign attackers
    for attacker in campaign_attackers:
        attacker_campaign_count[attacker] += 1
        attacker_profits[attacker] += profit
        
        attacker_bots[attacker].add(bot)
        bot_attackers[bot].add(attacker)
        
        attacker_suppressed_contracts[attacker].add(suppressed_contract)
        suppressed_contract_attackers[suppressed_contract].add(attacker)
    
# some numbers
print("number of campaigns", len(campaigns))
print("number of unique attackers involved", len(attackers))
print("number of unique suppressed contracts involved", len(suppressed_contracts))
print("number of unique bots involved", len(bots))

df_attackers = pd.DataFrame({
    "Attacker Hash": pd.Series([], dtype="object"),
    "Campaigns": pd.Series([], dtype="int"),
    "Rounds": pd.Series([], dtype="int"),
    "Transactions": pd.Series([], dtype="int"),
    "Bots": pd.Series([], dtype="int"),
    "Suppressed Contracts": pd.Series([], dtype="int"),
    "Profit": pd.Series([], dtype="float"),
    "Bot Cluster ID": pd.Series([], dtype="int"),
})

for attacker in attackers:
    df_attackers = df_attackers.append({
        "Attacker Hash": attacker,
        "Campaigns": attacker_campaign_count[attacker],
        "Rounds": attacker_round_count[attacker],
        "Transactions": attacker_transaction_count[attacker],
        "Bots": len(attacker_bots[attacker]),
        "Suppressed Contracts": len(attacker_suppressed_contracts[attacker]),
        "Profit": attacker_profits[attacker],
        "Bot Cluster ID": attacker_hash_to_cluster_id.get(attacker),  # will give None if bot clusters are not updated
    }, ignore_index=True)
    
df_attackers.to_csv(os.path.join(data_directory, "suppression_attackers.csv"), index=False)
    
df_attackers.sort_values(by="Profit", ascending=False)

df_bots = pd.DataFrame({
    "Bot Hash": pd.Series([], dtype="object"),
    "Campaigns": pd.Series([], dtype="int"),
    "Rounds": pd.Series([], dtype="int"),
    "Transactions": pd.Series([], dtype="int"),
    "Attackers": pd.Series([], dtype="int"),
    "Suppressed Contracts": pd.Series([], dtype="int"),
    "Profit": pd.Series([], dtype="float"),
    "Bot Cluster ID": pd.Series([], dtype="int"),
    "First Block": pd.Series([], dtype="int"),
})

for bot in bots:
    df_bots = df_bots.append({
        "Bot Hash": bot,
        "Campaigns": bot_campaign_count[bot],
        "Rounds": bot_round_count[bot],
        "Transactions": bot_transaction_count[bot],
        "Attackers": len(bot_attackers[bot]),
        "Suppressed Contracts": len(bot_suppressed_contracts[bot]),
        "Profit": bot_profits[bot],
        "Bot Cluster ID": bot_hash_to_cluster_id.get(bot),  # will give None if bot clusters are not updated
        "First Block": bot_first_blocks[bot],
    }, ignore_index=True)
    
df_bots.to_csv(os.path.join(data_directory, "suppression_bots.csv"), index=False)

df_bots.sort_values(by="Profit", ascending=False)

pd.concat([
    df_bots[df_bots["Bot Cluster ID"] == 0][["Bot Cluster ID", "Campaigns", "Rounds", "Profit"]],
    df_bots[df_bots["Bot Cluster ID"] > 0].groupby("Bot Cluster ID").agg({
        "Campaigns": "sum",
        "Rounds": "sum",
        "Profit": "sum",
    }).reset_index()
], axis=0).sort_values(by="Profit", ascending=False)

df_suppresed_contracts = pd.DataFrame({
    "Suppressed Contract Address": pd.Series([], dtype="object"),
    "Contract Name": pd.Series([], dtype="object"),
    "Attacks": pd.Series([], dtype="int"),
    "Rounds": pd.Series([], dtype="int"),
    "Transactions": pd.Series([], dtype="int"),
    "Attackers": pd.Series([], dtype="int"),
    "Bots": pd.Series([], dtype="int"),
    "Attacker Clusters": pd.Series([], dtype="int"),
})

for suppressed_contract in suppressed_contracts:
    df_suppresed_contracts = df_suppresed_contracts.append({
        "Suppressed Contract Address": suppressed_contract,
        "Contract Name": suppressed_contract_names[suppressed_contract],
        "Attacks": suppressed_contract_campaign_count[suppressed_contract],
        "Rounds": suppressed_contract_round_count[suppressed_contract],
        "Transactions": suppressed_contract_transaction_count[suppressed_contract],
        "Attackers": len(suppressed_contract_attackers[suppressed_contract]),
        "Bots": len(suppressed_contract_bots[suppressed_contract]),
        "Attacker Clusters": len({bot_hash_to_cluster_id.get(bot)  # will give None if bot clusters are not updated
                             for bot in suppressed_contract_bots[suppressed_contract]}),
    }, ignore_index=True)


def rename_contract_name(row):
    if row["Suppressed Contract Address"] == "0x5D0d76787D9d564061dD23f8209F804a3b8AD2F2":
        row = pd.Series(row)
        row["Contract Name"] = "Peach Will"
    return row
    
df_suppresed_contracts = df_suppresed_contracts.apply(rename_contract_name, axis=1)

df_suppresed_contracts = df_suppresed_contracts.sort_values(by=[
    "Attacks",
    "Transactions",
], ascending=False)

df_suppresed_contracts

## Save latex tables for the publication

with pd.option_context("display.max_colwidth", 10000):
    latex = df_suppresed_contracts.to_latex(index=False, escape=False, formatters={
        "Suppressed Contract Address": lambda value: "\\href{{https://etherscan.io/address/{}}}{{{}}}".format(value, value),
        "Contract Name": lambda value: value,
        "Attacks": str,
        "Rounds": str,
        "Transactions": str,
        "Attacker Accounts": str,
        "Bots Contracts": str,
        "Attacker Clusters": str,
    })

    with open(os.path.join(data_directory, "suppression_victim_stats.tex"), "w") as latex_file:
        latex_file.write(latex)
