==> scripts/analysis/suppression/5 - Analyze competition.ipynb <==
# Analyze competition

import json
import os

import matplotlib.pyplot as plt
from pandas.plotting import register_matplotlib_converters


register_matplotlib_converters()

%matplotlib inline

plt.style.use("seaborn")

data_directory = os.path.join("..", "..", "..", "data")

campaigns_file_path = os.path.join(data_directory, "suppression_campaigns.json")
if not os.path.exists(campaigns_file_path):
    with tarfile.open(os.path.join(data_directory, "suppression_campaigns.tar.xz"), "r", encoding="utf-8") as compressed_file:
        compressed_file.extract("suppression_campaigns.json", data_directory)

with open(campaigns_file_path, "r", encoding="utf-8") as campaigns_file:
    campaigns = []
    line = campaigns_file.readline().strip()
    while line != "":
        campaigns.append(json.loads(line))
        line = campaigns_file.readline().strip()

with open(os.path.join(data_directory, "suppression_bot_hash_to_cluster_id.json"), "r") as json_file:
    bot_hash_to_cluster_id = json.load(json_file)

attacks_by_victim = {}

for campaign in campaigns:
    bot = campaign["bot_address"]
    victim = campaign["suppressed_contract_address"]
    
    if victim not in attacks_by_victim:
        attacks_by_victim[victim] = []
    
    for round_ in campaign["rounds"]:
        first_block = round_["first_block"]
        number_of_blocks = round_["nr_of_blocks"]

        attacks_by_victim[victim].append({
            "bot_cluster_id": bot_hash_to_cluster_id[bot],
            "bot": bot,
            "first_block": round_["first_block"],
            "number_of_blocks": round_["nr_of_blocks"],
        })

competition_by_victim = {}

for victim in attacks_by_victim.keys():
    if len(attacks_by_victim[victim]) > 1:
        competition_by_victim[victim] = sorted(attacks_by_victim[victim], key=lambda attack: attack["first_block"])

for victim, attacks in competition_by_victim.items():
    attack = attacks[0]
    for next_attack in attacks[1:]:
        if attack["first_block"] + attack["number_of_blocks"] >= next_attack["first_block"]:
            print("victim:", victim)
            print("first attack:", attack)
            print("second attack:", next_attack)
            print("---")
        attack = next_attack
